Change aggregation to write to file directly and remove the WriteSet operator. This fixed the problem, though the root cause is still unclear. There maybe problems in WriteSet following an Aggregation Job Stage.

+++ src/queryExecution/headers/PartitionedHashSet.h
@@ -129,7 +129,7 @@ public:
 
     // get size
     size_t getSize() override {
-        return pageSize * partitionPages.size();
+        return pageSize * (size_t)(partitionPages.size());
     }
 };
 }
